pyproject.toml
[tool.poetry]
name = "homework-drf"
version = "0.1.0"
description = ""
authors = ["Aleksandr <aqwa06@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
pillow = "^11.0.0"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.6.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
